Grid.hpp: use u64 instead of size_t

+++ src/samarium/graphics/Grid.hpp
@@ -150,7 +150,7 @@ template <typename T> class Grid
     auto max_size() const { return this->data.size(); } // for stl compatibility
     auto empty() const { return this->data.size() == 0; }
 
-    auto bounding_box() const { return BoundingBox<size_t>{Indices{}, dims - Indices{1, 1}}; }
+    auto bounding_box() const { return BoundingBox<u64>{Indices{}, dims - Indices{1, 1}}; }
 
     auto fill(const T& value) { this->data.fill(value); }
